+++ source/texk/web2c/luatexdir/pdf/pdfpage.c
@@ -124,20 +124,20 @@ boolean calc_pdfpos(pdfstructure * p, scaledpos pos)
     case PMODE_CHARARRAY:
         switch (p->wmode) {
         case WMODE_H:
-            new.h = lround((pos.h * p->k1 - p->pdf_tj_pos.h.m) * p->k2);
+            new.h = lround((pos.h * p->k1 - (double)p->pdf_tj_pos.h.m) * p->k2);
             new.v = lround(pos.v * p->k1);
             p->tj_delta.m =
-                -lround((new.h - p->cw.m) / ten_pow[p->cw.e - p->tj_delta.e]);
+                -lround((double)((new.h - p->cw.m) / ten_pow[p->cw.e - p->tj_delta.e]));
             p->tm[5].m = new.v - p->pdf_bt_pos.v.m;     /* p->tm[4] is meaningless */
             if (p->tj_delta.m != 0 || new.v != p->pdf.v.m)
                 move_pdfpos = TRUE;
             break;
         case WMODE_V:
             new.h = lround(pos.h * p->k1);
-            new.v = lround((p->pdf_tj_pos.v.m - pos.v * p->k1) * p->k2);
+            new.v = lround(((double)p->pdf_tj_pos.v.m - pos.v * p->k1) * p->k2);
             p->tm[4].m = new.h - p->pdf_bt_pos.h.m;     /* p->tm[5] is meaningless */
             p->tj_delta.m =
-                -lround((new.v - p->cw.m) / ten_pow[p->cw.e - p->tj_delta.e]);
+                -lround((double)((new.v - p->cw.m) / ten_pow[p->cw.e - p->tj_delta.e]));
             if (p->tj_delta.m != 0 || new.h != p->pdf.h.m)
                 move_pdfpos = TRUE;
             break;
